+++ src/main/java/tallestred/piglinproliferation/common/entities/ai/behaviors/ThrowPotionAtSelfTask.java
@@ -47,7 +47,7 @@ public class ThrowPotionAtSelfTask<E extends PiglinAlchemist> extends BaseThrowP
 
     @Override
     protected void tick(ServerLevel level, E alchemist, long gameTIme) {
-        BlockPos lookPos = new BlockPos(alchemist.position().x, alchemist.getBoundingBox().minY - 0.5000001D, alchemist.position().z);
+        BlockPos lookPos = new BlockPos((int) alchemist.position().x, (int) (alchemist.getBoundingBox().minY - 0.5000001D), (int) alchemist.position().z);
         alchemist.getBrain().setMemory(MemoryModuleType.LOOK_TARGET, new BlockPosTracker(lookPos));
         List<LivingEntity> list = alchemist.level.getEntitiesOfClass(LivingEntity.class, alchemist.getBoundingBox().inflate(5.0D, 3.0D, 5.0D));
         if (!list.isEmpty()) {
